fix args name

+++ wiki_poc/models/runners/abstract_runner.py
@@ -107,7 +107,7 @@ class AbstractRunner():
         model_path = self.names()[self.model_name]
         # if GPU is available, load in 8bit mode
         if torch.cuda.is_available():
-            return self._model_loader().from_pretrained(model_path, load_in_8bit=True, device_map="auto")
+            return self._model_loader().from_pretrained(model_path, load_in_8bit=True, torch_dtype=torch.float16, device_map="auto")
         else:
             logging.warning("GPU not available, loading model in FP32 mode on CPU. This will be very slow.")
             return self._model_loader().from_pretrained(model_path)
